config-manager/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
# 数据库驱动
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyyaml==6.0.1
cryptography>=41.0.0
redis==5.0.1
prometheus-client==0.19.0
psutil==5.9.6
aiohttp==3.9.1
asyncio-throttle==1.0.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
email-validator==2.1.0